Add resolver backed by hosts file
