+++ dres.go
@@ -17,6 +17,62 @@ type Dres struct {
 	Resolvers map[string][]Resolver
 }
 
+func Load(config Config) Dres {
+	var resolverByName = make(map[string]Resolver)
+
+	for resolverName, resolverConfig := range config.Resolvers {
+		resolver, err := LoadResolver(resolverName, resolverConfig)
+		if err != nil {
+			log.Fatalf("Error constructing resolver: %s", err)
+		} else {
+			resolverByName[resolverName] = resolver
+			log.Printf("Loaded resolver %s of type %s", resolverName, resolverConfig.Type)
+		}
+	}
+
+	var networks []Network
+	for rangeName, cidr := range config.CIDRS {
+		_, ipNet, err := net.ParseCIDR(cidr)
+		if err != nil {
+			log.Fatalf("Error reading CIDR: %s", err)
+		} else {
+			network := Network{
+				Name: rangeName,
+				Net:  *ipNet,
+			}
+			networks = append(networks, network)
+			log.Printf("Loaded network %s: %s", rangeName, cidr)
+		}
+	}
+
+	var resolvers = make(map[string][]Resolver)
+	for rangeName, resolverNames := range config.Configuration {
+		log.Printf("Range %s has following resolvers:", rangeName)
+		_, ok := resolvers[rangeName]
+		if !ok {
+			resolvers[rangeName] = make([]Resolver, len(resolverNames))
+		}
+		for i, resolverName := range resolverNames {
+			resolvers[rangeName][i] = resolverByName[resolverName]
+			log.Printf(" - %s", resolverName)
+		}
+	}
+
+	return Dres{
+		Networks:  networks,
+		Resolvers: resolvers,
+	}
+}
+
+func (dres Dres) GetResolvers(addr net.Addr) []Resolver {
+	networkName, err := dres.GetNetworkName(addr)
+	if err != nil {
+		log.Printf("Network for address %s not found: %s", addr, err)
+		return make([]Resolver, 0)
+	}
+	return dres.Resolvers[networkName]
+}
+
 func (dres Dres) HandleFunc(writer dns.ResponseWriter, msg *dns.Msg) {
 	log.Printf("Request from %s", writer.RemoteAddr())
 	for _, question := range msg.Question {
@@ -40,24 +96,6 @@ func (dres Dres) HandleFunc(writer dns.ResponseWriter, msg *dns.Msg) {
 	_ = writer.Close()
 }
 
-func (dres Dres) GetResolvers(addr net.Addr) []Resolver {
-	networkName, err := dres.GetNetworkName(addr)
-	if err != nil {
-		log.Printf("Network for address %s not found: %s", addr, err)
-		return make([]Resolver, 0)
-	}
-	return dres.Resolvers[networkName]
-}
-
-func GetIP(addr net.Addr) net.IP {
-	switch addr := addr.(type) {
-	case *net.UDPAddr:
-		return addr.IP
-	case *net.TCPAddr:
-		return addr.IP
-	}
-	return nil
-}
 func (dres Dres) GetNetworkName(addr net.Addr) (string, error) {
 	for _, network := range dres.Networks {
 		if network.Net.Contains(GetIP(addr)) {
@@ -69,13 +107,9 @@ func (dres Dres) GetNetworkName(addr net.Addr) (string, error) {
 }
 
 func main() {
-	config := LoadConfig()
-	dres := LoadDres(config)
+	dres := Load(LoadConfig())
 
-	server := dns.Server{
-		Addr: ":53",
-		Net:  "udp",
-	}
+	server := dns.Server{Addr: ":53", Net: "udp"}
 
 	dns.HandleFunc(".", dres.HandleFunc)
