+++ hosts.file
@@ -1,0 +1,6 @@
+# comment 1
+
+0.0.0.0 ad-server-1.com
+
+# comment 2
+0.0.0.0 ad-server-2.com
